Repositori Data/Latihan  Menyimpan Data ke Google Sheets/requirements.txt
google-api-python-client~=2.152
google-auth
google-auth-httplib2
google-auth-oauthlib
